backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic[email]>=2.4.2
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
alembic>=1.12.1
# For PostgreSQL support (when switching from SQLite)
psycopg2-binary>=2.9.9
# For testing
pytest>=7.4.3
httpx>=0.25.1
